pyproject.toml
[project]
name = "operag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "numpy>=2.3.3",
    "path>=17.1.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.1",
]
